third_party/blink/web_tests/virtual/css-anchor-update-enabled/external/wpt/css/css-anchor-position/anchor-scroll-position-try-001-expected.txt

This is a testharness.js-based test.
[FAIL] Scroll down until the top edge of #anchor touches container but not overflowing
  assert_equals: Anchored element should be at the top of anchor expected 200 but got 300
[FAIL] Scroll further up, where the second option no longer fits
  assert_equals: Anchored element should be at the top of anchor expected 400 but got 500
Harness: the test ran to completion.
